requirements.txt
requests==2.28.1
numpy==1.23.1
pandas~=1.5.0
xlwings~=0.27.15
future>=0.18.2
tkcalendar>=1.6.1
ttkthemes>=3.2.2
selenium~=4.4.3
gspread~=5.5.0
oauth2client>=4.1.3
google-api-python-client~=2.62.0
df2gspread>=1.0.4
webdriver_manager==3.8.2
altgraph~=0.17.3
arrow~=1.2.3
async-generator==1.10
attrs~=22.1.0
Babel==2.10.3
binaryornot==0.4.4
botcity-framework-base==0.3.1
botcity-framework-core==0.2.9
botcity-maestro-sdk==0.1.5
certifi~=2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
configparser~=5.3.0
cookiecutter==2.1.1
crayons==0.4.0
cryptography~=38.0.1
DateTime~=4.7
distlib==0.3.5
et-xmlfile==1.1.0
h11~=0.14.0
idna~=3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
MouseInfo==0.1.3
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
pefile==2022.5.30
Pillow==9.2.0
PyAutoGUI==0.9.53
pybrowsers==0.5.1
pycparser==2.21
PyGetWindow==0.0.9
pyglet~=1.5.27
pyinstaller~=5.4.1
pyinstaller-hooks-contrib==2022.8
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv~=0.21.0
python-slugify==6.1.2
python-xlib==0.31
pytweening==1.0.4
pytz~=2022.2.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
requests_download==0.1.2
six==1.16.0
sniffio~=1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.21.0
trio-websocket==0.9.2
urllib3~=1.26.12
yarg==0.1.9
zope.interface==5.4.0
pip~=23.1.2
argparse~=1.4.0
wheel~=0.37.1
rsa~=4.9
pyasn1~=0.4.8
gast~=0.4.0
h5py~=3.8.0
keras~=2.11.0
setuptools~=60.2.0
tensorflow~=2.11.0
scipy~=1.10.1
wrapt~=1.14.1
cachetools~=5.2.0
joblib~=1.2.0
pdf2image~=1.16.3
fillpdf~=0.7.2
scikit-learn~=1.2.1
threadpoolctl~=3.1.0
Markdown~=3.4.1
Werkzeug~=2.2.3
pyparsing~=3.0.9
httplib2~=0.20.4
oauthlib~=3.2.1
termcolor~=2.2.0
darkdetect~=0.8.0
tensorboard~=2.11.2
flatbuffers~=23.1.21
packaging~=23.0
astunparse~=1.6.3
uritemplate~=4.1.1
customtkinter~=5.1.3
wsproto~=1.2.0
tqdm~=4.64.1
imutils~=0.5.4